requirements.txt
gspread==5.7.2
oauth2client==4.1.3
beautifulsoup4==4.11.2
selenium==4.8.2
python-dotenv==1.0.0
